Pagination code refactoring
