+++ components/Pagination.js
@@ -7,12 +7,12 @@ export default function Pagination({ totalPages, currentPage }) {
   return (
     <nav className="flex justify-between">
       {!prevPage && (
-        <button rel="previous" className="cursor-auto">
-          <strike>Previous</strike>
+        <button rel="previous" className="cursor-auto disabled:opacity-50" disabled={!prevPage}>
+          Previous
         </button>
       )}
       {prevPage && (
-        <Link href={currentPage - 1 === 1 ? `/blog/` : `/blog/page/${currentPage - 1}`}>
+        <Link href={currentPage - 1 === 1 ? `/blog/` : `/blog/${currentPage - 1}`}>
           <button rel="previous">Previous</button>
         </Link>
       )}
@@ -20,12 +20,12 @@ export default function Pagination({ totalPages, currentPage }) {
         {currentPage} of {totalPages}
       </span>
       {!nextPage && (
-        <button rel="next" className="cursor-auto">
-          <strike>Next</strike>
+        <button rel="next" className="cursor-auto disabled:opacity-50" disabled={!nextPage}>
+          Next
         </button>
       )}
       {nextPage && (
-        <Link href={`/blog/page/${currentPage + 1}`}>
+        <Link href={`/blog/${currentPage + 1}`}>
           <button rel="next">Next</button>
         </Link>
       )}
